requirements.txt
streamlit
pandas
pyairtable
python-dotenv
bcrypt
streamlit-authenticator
plotly
